handled error: error: comparison of integer expressions of different signedness: ‘long unsigned int’ and ‘int’ [-Werror=sign-compare]

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -17,7 +17,7 @@ int _pow(int x, int y)
  */
 void print_binary(unsigned long int n)
 {
-	int num = 0, start, bg;
+	unsigned long int num = 0, start, bg;
 
 	if (n == 0)
 	{
